requirements.txt
astropy==4.3.1
ccdproc==2.4.3
cupy==13.3.0
ipykernel==6.29.5
matplotlib==3.10.0
numba==0.61.0
numpy==1.21.6
psutil==7.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
Requests==2.32.3
tqdm==4.67.1
watchdog==6.0.0
